Module type Ox.OXMODULE

OXMODULE is the type of modules that implement Ox controller callback functions. Modules of type OXMODULE are passed to the Make functor, producing an Ox controller.

val switch_connected : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.SwitchFeatures.t -> unit

switch_connected sw is a callback invoked with sw when a switch with identifer sw connects to the controller.

val switch_disconnected : Frenetic_kernel.OpenFlow0x01.switchId -> unit

switch_disconnected sw is a callback invoked with sw when a switch with identifer sw disconnects from the controller.

val packet_in : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.xid -> Frenetic_kernel.OpenFlow0x01.packetIn -> unit

packet_in sw xid pkt is a callback invoked when a packet pkt with transaction ID xid from switch sw arrives at the controller.

val barrier_reply : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.xid -> unit

barrier_reply sw xid is a callback invoked when a barrier reply with transaction ID xid from switch sw arrives at the controller.

val stats_reply : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.xid -> Frenetic_kernel.OpenFlow0x01.reply -> unit

stats_reply sw xid rep is a callback invoked when switch sw responds with a reply rep to a statistics request with transaction ID xid.

val cleanup : unit -> unit

cleanup is called when an exception stops the running of the main controller loop.