Module Netkat.Semantics

Frenetic_kernel.Packets

If we only defined the semantics and were not building a system, a packet would only be a record of headers. However, the runtime needs to apply eval to packets contained in PACKET_IN mesages. For the runtime, packets also carry a payload that is unmodified by eval.

module HeadersValues : sig ... end

A map keyed by header names.

type packet = {

switch : Frenetic_netkat.Syntax.switchId;

headers : HeadersValues.t;

payload : Frenetic_netkat.Syntax.payload;

}

module PacketSet : Core.Set.S with type PacketSet.Elt.t = packet

Semantics

val eval : packet -> Frenetic_netkat.Syntax.policy -> PacketSet.t

val eval_pipes : packet -> Frenetic_netkat.Syntax.policy -> (string * packet) list * (string * packet) list * packet list

val size : Frenetic_netkat.Syntax.policy -> int

val queries_of_policy : Frenetic_netkat.Syntax.policy -> string list

val switches_of_policy : Frenetic_netkat.Syntax.policy -> Frenetic_netkat.Syntax.switchId list