Module Ox.DefaultHandlers

Provides default implementations for some advanced event handlers, reducing clutter in simple controllers.

These event handlers simply ignore the messages they receive.

val switch_connected : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.SwitchFeatures.t -> unit

val switch_disconnected : Frenetic_kernel.OpenFlow0x01.switchId -> unit

val barrier_reply : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.xid -> unit

val stats_reply : Frenetic_kernel.OpenFlow0x01.switchId -> Frenetic_kernel.OpenFlow0x01.xid -> Frenetic_kernel.OpenFlow0x01.reply -> unit

val cleanup : unit -> unit