Module Ox.Make

Given an OXMODULE module, build an Ox controller that, listens on port 6633 for messages from OpenFlow-enabled switches. Messages and network events are passed to the appropriate OXMODULE callbacks.

Parameters

OxModule : OXMODULE

Signature

val start : unit -> unit

start is called to start the Ox controller. Note that execution will not be transfered back to the calling thread.