Module Frenetic_async__Ox

module DefaultHandlers : sig ... end

Provides default implementations for some advanced event handlers, reducing clutter in simple controllers.

module type OXMODULE = sig ... end

OXMODULE is the type of modules that implement Ox controller callback functions. Modules of type OXMODULE are passed to the Make functor, producing an Ox controller.

module Platform : sig ... end

Platform defines functions for sending OpenFlow messages to switches and wrapping thunks with timeouts.

module Make : functor (OxModule : OXMODULE) -> sig ... end

Given an OXMODULE module, build an Ox controller that, listens on port 6633 for messages from OpenFlow-enabled switches. Messages and network events are passed to the appropriate OXMODULE callbacks.