Module OpenFlow0x01.Action

type sequence = action list

val move_controller_last : sequence -> sequence

move_controller_last seq produces a semantically-equivalent list of actions with actions that send packets to the controller moved to the end. This works around a known bug in the OpenFlow reference switch where actions in an action sequence after a "send to controller" (Output (Controller n)) action are ignored.

val to_string : action -> string

val sequence_to_string : sequence -> string

val marshal : action -> Cstruct.t -> int

val parse : Cstruct.t -> action

val size_of : action -> int