Module Frenetic_netkat.Global_compiler

module FDD : module type of Local_compiler.FDD

module Automaton : sig ... end

Intermediate representation of global compiler: NetKAT Automata

val compile : ?⁠options:Local_compiler.compiler_options -> ?⁠pc:Fdd.Field.t -> ?⁠ing:Syntax.pred -> Syntax.policy -> FDD.t

compile p compiles the policy p into an FDD. The pc field is used for internal bookkeeping and must *not* be accessed or written to by the input policy p.